requirements.txt
typer[all]==0.12.3
rich==13.7.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
redis==5.0.1
python-dotenv==1.0.0
pypdf==4.0.1
tiktoken==0.7.0

